[refactor] About 페이지 WorkProcess 컴포넌트 분리
- WorkProcess와 WorkProcessItem 컴포넌트로 분리
