+++ front/src/components/UI/molecules/WorkProcessItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './WorkProcessItem';
