+++ front/src/components/UI/organisms/WorkProcess/index.ts
@@ -1,0 +1,1 @@
+export { default } from './WorkProcess';
