+++ front/src/routes/About/styled.ts
@@ -26,42 +26,6 @@ export const AboutItemWrapper = styled.div`
   }
 `;
 
-export const WorkProcessWrapper = styled.div`
-  margin: 40px 0;
-  display: flex;
-  justify-content: space-around;
-  align-items: flex-end;
-  flex-wrap: wrap;
-`;
-
-export const WorkProcessItemWrapper = styled.div`
-  width: 100px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  & > img {
-    width: 100px;
-  }
-
-  & > span {
-    margin-top: 14px;
-    text-align: center;
-    font-size: 22px;
-    font-weight: 800;
-  }
-
-  @media (max-width: ${({ theme }) => theme.BP.TABLET}) {
-    & > img {
-      width: 60px;
-    }
-
-    & > span {
-      font-size: 18px;
-    }
-  }
-`;
-
 export const SkillListWrapper = styled.div`
   margin: 30px 0;
   display: grid;
